+++ static/app/views/preprod/buildComparison/main/sizeCompareSelectionContent.tsx
@@ -35,6 +35,8 @@ import {
 } from 'sentry/views/preprod/types/buildDetailsTypes';
 import type {ListBuildsApiResponse} from 'sentry/views/preprod/types/listBuildsTypes';
 
+import {SizeCompareSelectedBuilds} from './sizeCompareSelectedBuilds';
+
 interface SizeCompareSelectionContentProps {
   headBuildDetails: BuildDetailsApiResponse;
   baseBuildDetails?: BuildDetailsApiResponse;
@@ -55,14 +57,6 @@ export function SizeCompareSelectionContent({
   const [selectedBaseBuild, setSelectedBaseBuild] = useState<
     BuildDetailsApiResponse | undefined
   >(baseBuildDetails);
-
-  const headPrNumber = headBuildDetails.vcs_info?.pr_number;
-  const headSha = headBuildDetails.vcs_info?.head_sha?.substring(0, 7);
-  const headBranchName = headBuildDetails.vcs_info?.head_ref;
-
-  const basePrNumber = selectedBaseBuild?.vcs_info?.pr_number;
-  const baseSha = selectedBaseBuild?.vcs_info?.head_sha?.substring(0, 7);
-  const baseBranchName = selectedBaseBuild?.vcs_info?.head_ref;
 
   const queryParams: Record<string, any> = {
     per_page: 25,
@@ -110,90 +104,22 @@ export function SizeCompareSelectionContent({
 
   return (
     <Stack gap="xl">
-      <Flex align="center" gap="lg" width="100%" justify="center">
-        <Flex align="center" gap="sm">
-          <IconLock size="xs" locked />
-          <Text bold>{t('Your build:')}</Text>
-          <Flex align="center" gap="md">
-            <Text size="sm" variant="accent" bold>
-              {headPrNumber && `#${headPrNumber} `}
-            </Text>
-            <Text size="sm" variant="accent" bold>
-              {headSha && (
-                <Flex align="center" gap="xs">
-                  <IconCommit size="xs" />
-                  {headSha}
-                </Flex>
-              )}
-            </Text>
-          </Flex>
-          <BuildBranch>
-            <Text size="sm" variant="muted">
-              {headBranchName}
-            </Text>
-          </BuildBranch>
-        </Flex>
+      <SizeCompareSelectedBuilds
+        isComparing={isComparing}
+        headBuildDetails={headBuildDetails}
+        baseBuildDetails={selectedBaseBuild}
+        onClearBaseBuild={() => setSelectedBaseBuild(undefined)}
+        onTriggerComparison={() => {
+          if (!selectedBaseBuild) {
+            return;
+          }
 
-        <Text>{t('vs')}</Text>
-
-        <Flex align="center" gap="sm">
-          {selectedBaseBuild ? (
-            <SelectedBaseBuild align="center" gap="sm">
-              <IconFocus size="xs" color="purple400" />
-              <Text size="sm" variant="accent" bold>
-                {t('Comparison:')}
-              </Text>
-              <Text size="sm" variant="accent" bold>
-                {basePrNumber && `#${basePrNumber} `}
-                {baseSha && (
-                  <Flex align="center" gap="xs">
-                    <IconCommit size="xs" color="purple400" />
-                    {baseSha}
-                  </Flex>
-                )}
-              </Text>
-              <BaseBuildBranch>
-                <Text size="sm" variant="muted">
-                  {baseBranchName}
-                </Text>
-              </BaseBuildBranch>
-              <Button
-                onClick={e => {
-                  e.stopPropagation();
-                  setSelectedBaseBuild(undefined);
-                }}
-                size="zero"
-                priority="transparent"
-                borderless
-                aria-label={t('Clear base build')}
-                icon={<IconClose size="xs" color="purple400" />}
-              />
-            </SelectedBaseBuild>
-          ) : (
-            <SelectBuild>
-              <Text size="sm">{t('Select a build')}</Text>
-            </SelectBuild>
-          )}
-        </Flex>
-
-        <Flex align="center" gap="sm">
-          <Button
-            onClick={() => {
-              if (selectedBaseBuild) {
-                triggerComparison({
-                  baseArtifactId: selectedBaseBuild.id.toString(),
-                  headArtifactId: headBuildDetails.id.toString(),
-                });
-              }
-            }}
-            disabled={!selectedBaseBuild || isComparing}
-            priority="primary"
-            icon={<IconTelescope size="sm" />}
-          >
-            {isComparing ? t('Comparing...') : t('Compare builds')}
-          </Button>
-        </Flex>
-      </Flex>
+          triggerComparison({
+            baseArtifactId: selectedBaseBuild.id.toString(),
+            headArtifactId: headBuildDetails.id.toString(),
+          });
+        }}
+      />
 
       {buildsQuery.isLoading && <LoadingIndicator />}
       {buildsQuery.isError && <Alert type="error">{buildsQuery.error?.message}</Alert>}
@@ -218,32 +144,6 @@ export function SizeCompareSelectionContent({
     </Stack>
   );
 }
-
-const BuildBranch = styled('span')`
-  padding: ${p => p.theme.space.xs} ${p => p.theme.space.sm};
-  background-color: ${p => p.theme.gray100};
-  border-radius: ${p => p.theme.borderRadius};
-`;
-
-const BaseBuildBranch = styled('span')`
-  padding: ${p => p.theme.space['2xs']} ${p => p.theme.space.sm};
-  background-color: ${p => p.theme.gray100};
-  border-radius: ${p => p.theme.borderRadius};
-`;
-
-const SelectBuild = styled('div')`
-  border: 1px solid ${p => p.theme.border};
-  border-radius: ${p => p.theme.borderRadius};
-  border-style: dashed;
-  padding: ${p => p.theme.space.md} ${p => p.theme.space.lg};
-`;
-
-const SelectedBaseBuild = styled(Flex)`
-  background-color: ${p => p.theme.surface100};
-  border: 1px solid ${p => p.theme.focusBorder};
-  border-radius: ${p => p.theme.borderRadius};
-  padding: ${p => p.theme.space.sm} ${p => p.theme.space.md};
-`;
 
 interface BuildItemProps {
   build: BuildDetailsApiResponse;
